+++ views/expenses/search.ejs
@@ -94,8 +94,14 @@
                     </div>
                 </div>
             </div>
-            <div class="d-grid col-12 mx-auto">
+            <div class="d-grid col-12 mx-auto my-3">
                 <button class="btn btn-outline-success">Išči</button>
+            </div>
+        </form>
+        <form action="/search" method="POST">
+            <input type="hidden" name="lastMonth" value="true">
+            <div class="d-grid col-12 mx-auto my-3">
+                <button class="btn btn-outline-success">Pretekli mesec</button>
             </div>
         </form>
     </div>
